refactor: relocate Formulas/ into src/game to fix layer cycle
formulas.ts was pure domain logic living under src/app/ while src/game/
depended back on it, creating an upward-then-downward import cycle
flagged in docs/technical-debt.md (A2).
