+++ src/app/TombLevel/ComparePuzzle.tsx
@@ -7,7 +7,7 @@ import { NumberChest } from "@/ui/molecules/NumberChest"
 import crocodileOpen from "@/assets/crocodile-250.png"
 import crocodileClosed from "@/assets/crocodile-closed-250.png"
 import clsx from "clsx"
-import { formulaToString } from "@/app/Formulas/formulas"
+import { formulaToString } from "@/game/formulas/formulas"
 import { useCrocodilePuzzleControls } from "./useComparePuzzleControls"
 import { use, useEffect, type FC, type ReactNode } from "react"
 import { FezContext } from "../fez/context"
