+++ src/game/formulas/filledPositions.ts
