+++ src/game/formulas/formulas.spec.ts
