+++ src/ui/organisms/TombTableau.stories.tsx
@@ -5,7 +5,7 @@ import { hashString } from "@/support/hashString"
 import { generateRewardCalculation } from "@/game/puzzles/tableau/generateRewardCalculation"
 import { useMemo } from "react"
 import { TombTableau } from "./TombTableau"
-import { createPositionOverview } from "@/app/Formulas/filledPositions"
+import { createPositionOverview } from "@/game/formulas/filledPositions"
 import { useTableauTranslations } from "@/data/useTableauTranslations"
 import { resolveHieroglyphSymbol } from "@/data/resolveHieroglyphSymbol"
